# frontend/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Frontend",
        "version": "v0.0.0"
    },
    "paths": {
        "/api/frontend/v1/oidc/start-auth": {
            "get": {
                "tags": [
                    "oidc"
                ],
                "summary": "Handler for OIDC's login endpoint",
                "description": "Handler for OIDC's login endpoint",
                "operationId": "start_auth",
                "responses": {
                    "400": {
                        "description": "Client side error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server side error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "3XX": {
                        "description": "A generic http redirect",
                        "headers": {
                            "Location": {
                                "style": "simple",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/frontend/v1/oidc/finish-auth": {
            "get": {
                "tags": [
                    "oidc"
                ],
                "summary": "Handler for the OIDC endpoint the user will be redirected to from the OIDC provider",
                "description": "Handler for the OIDC endpoint the user will be redirected to from the OIDC provider",
                "operationId": "finish_auth",
                "parameters": [
                    {
                        "in": "query",
                        "name": "code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "style": "form"
                    },
                    {
                        "in": "query",
                        "name": "state",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "style": "form"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Client side error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server side error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "3XX": {
                        "description": "A generic http redirect",
                        "headers": {
                            "Location": {
                                "style": "simple",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/frontend/v1/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Use the local authentication for logging in",
                "description": "Use the local authentication for logging in",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "400": {
                        "description": "Client side error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server side error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Empty body"
                    }
                }
            }
        },
        "/api/frontend/v1/auth/logout": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Drop the current session and logg-out",
                "description": "Drop the current session and logg-out",
                "operationId": "logout",
                "responses": {
                    "400": {
                        "description": "Client side error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server side error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Empty body"
                    }
                }
            }
        },
        "/api/frontend/v1/ws": {
            "get": {
                "tags": [
                    "websocket"
                ],
                "operationId": "websocket",
                "responses": {
                    "101": {
                        "description": "Switching protocols",
                        "content": {
                            "application/octet-stream": {}
                        }
                    }
                }
            }
        },
        "/api/frontend/v1/users/me": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Retrieve the currently logged-in user",
                "description": "Retrieve the currently logged-in user",
                "operationId": "get_me",
                "responses": {
                    "400": {
                        "description": "Client side error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server side error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FullUser"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ApiErrorResponse": {
                "description": "The response that is sent in a case of an error",
                "type": "object",
                "properties": {
                    "message": {
                        "description": "A human-readable error message.\n\nMay be used for displaying purposes",
                        "type": "string"
                    },
                    "status_code": {
                        "$ref": "#/components/schemas/ApiStatusCode"
                    }
                },
                "required": [
                    "message",
                    "status_code"
                ]
            },
            "ApiStatusCode": {
                "description": "The Status code that are returned throughout the API",
                "type": "integer",
                "enum": [
                    1000,
                    2000
                ]
            },
            "FullUser": {
                "description": "The full representation for the user",
                "type": "object",
                "properties": {
                    "display_name": {
                        "description": "Used for displaying purposes",
                        "type": "string"
                    },
                    "uuid": {
                        "description": "The identifier of the user",
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "required": [
                    "display_name",
                    "uuid"
                ]
            },
            "LoginRequest": {
                "description": "The request for local authentication",
                "type": "object",
                "properties": {
                    "password": {
                        "description": "The password for the user",
                        "type": "string"
                    },
                    "username": {
                        "description": "The username that is used for logging in",
                        "type": "string"
                    }
                },
                "required": [
                    "password",
                    "username"
                ]
            },
            "WsClientMsg": {
                "description": "Websocket messages that originate from the client",
                "oneOf": []
            },
            "WsServerMsg": {
                "description": "Websocket messages that originate from the server",
                "oneOf": []
            }
        }
    }
}